pyproject.toml
[tool.poetry]
name = "python-coding-test"
version = "0.1.0"
description = ""
authors = ["Erdem Ozkol <github@erdemozkol.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
black = "^24.4.2"
pydantic-settings = "^2.2.1"
pydantic = "^2.7.1"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
pytest = "^8.2.1"
pytest-cov = "^5.0.0"
pre-commit = "^3.7.1"

[tool.black]
line-length = 100
target-version = ["py310"]


[tool.pytest.ini_options]
addopts = "--cov=src --cov-report=term-missing --cov-report=html"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
